+++ src/Api/ConfigInterface.php
@@ -11,6 +11,11 @@ use Magento\Framework\Exception\LocalizedException;
 interface ConfigInterface
 {
     public const CONFIG_PATH_CUSTOM_FEES = 'sales/custom_order_fees/custom_fees';
+    public const CONFIG_PATH_REPORTS_CUSTOM_ORDER_FEES_ENABLE_AGGREGATION
+        = 'reports/custom_order_fees/enable_aggregation';
+    public const CONFIG_PATH_REPORTS_CUSTOM_ORDER_FEES_AGGREGATION_TIME = 'reports/custom_order_fees/aggregation_time';
+    public const CONFIG_PATH_REPORTS_CUSTOM_ORDER_FEES_AGGREGATION_FREQUENCY
+        = 'reports/custom_order_fees/aggregation_frequency';
 
     /**
      * @return array{
@@ -40,4 +45,16 @@ interface ConfigInterface
      * @throws LocalizedException
      */
     public function getCustomFees(int|string|null $storeId = null): array;
+
+    public function isCustomOrderFeesReportAggregationEnabled(): bool;
+
+    /**
+     * @phpstan-return string Hour, minute and seconds separated by commas, e.g. "05,30,10"
+     */
+    public function getCustomOrderFeesReportAggregationTime(): string;
+
+    /**
+     * @phpstan-return 'D'|'W'|'M' "D" for daily, "W" for weekly or "M" for monthly
+     */
+    public function getCustomOrderFeesReportAggregationFrequency(): string;
 }